VZ-10281: Refactor VZ code and add analyze of Rancher clusterrepo objects (#6502)
