+++ tools/vz/pkg/analysis/internal/util/cluster/rancher.go
@@ -4,42 +4,10 @@
 package cluster
 
 import (
-	"encoding/json"
-	"fmt"
-	"io"
-	"os"
-
-	"github.com/verrazzano/verrazzano/tools/vz/pkg/analysis/internal/util/files"
+	"github.com/verrazzano/verrazzano/tools/vz/pkg/analysis/internal/util/cluster/rancher"
 	"github.com/verrazzano/verrazzano/tools/vz/pkg/analysis/internal/util/report"
 	"go.uber.org/zap"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-// Minimal definition of Rancher cluster object that only contains the fields that will be analyzed
-type rancherClusterList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []rancherCluster `json:"items"`
-}
-type rancherCluster struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	Spec              rancherClusterSpec   `json:"spec"`
-	Status            rancherClusterStatus `json:"status,omitempty"`
-}
-type rancherClusterSpec struct {
-	DisplayName string `json:"displayName,omitempty"`
-}
-type rancherClusterStatus struct {
-	Conditions []clusterCondition `json:"conditions,omitempty"`
-}
-type clusterCondition struct {
-	Status  corev1.ConditionStatus `json:"status"`
-	Type    string                 `json:"type"`
-	Reason  string                 `json:"reason,omitempty"`
-	Message string                 `json:"message,omitempty"`
-}
 
 // AnalyzeRancher handles the checking of the status of Rancher resources.
 func AnalyzeRancher(log *zap.SugaredLogger, clusterRoot string) error {
@@ -49,110 +17,6 @@ func AnalyzeRancher(log *zap.SugaredLogger, clusterRoot string) error {
 		PendingIssues: make(map[string]report.Issue),
 	}
 
-	return analyzeRancherClusters(log, clusterRoot, &issueReporter)
-}
-
-// analyzeRancherClusters - analyze the status of Rancher clusters
-func analyzeRancherClusters(log *zap.SugaredLogger, clusterRoot string, issueReporter *report.IssueReporter) error {
-	clusterPath := files.FindFileInClusterRoot(clusterRoot, "cluster.management.cattle.io.json")
-
-	// Parse the json into local struct
-	file, err := os.Open(clusterPath)
-	if err != nil {
-		// The file may not exist if Rancher is not installed.
-		log.Debugf("file %s not found", clusterPath)
-		return nil
-	}
-	defer file.Close()
-
-	fileBytes, err := io.ReadAll(file)
-	if err != nil {
-		log.Debugf("Failed reading Json file %s", clusterPath)
-		return err
-	}
-	clusterList := &rancherClusterList{}
-	err = json.Unmarshal(fileBytes, &clusterList)
-	if err != nil {
-		log.Debugf("Failed to unmarshal Rancher Cluster list at %s", clusterPath)
-		return err
-	}
-
-	for _, cluster := range clusterList.Items {
-		err = analyzeRancherCluster(clusterRoot, cluster, issueReporter)
-		if err != nil {
-			return err
-		}
-	}
-
-	issueReporter.Contribute(log, clusterRoot)
-	return nil
-}
-
-// analyzeRancherCluster - analyze a single Rancher cluster and report any issues
-func analyzeRancherCluster(clusterRoot string, cluster rancherCluster, issueReporter *report.IssueReporter) error {
-
-	var messages []string
-	var subMessage string
-	for _, condition := range cluster.Status.Conditions {
-		if condition.Status != corev1.ConditionTrue {
-			switch condition.Type {
-			case "Ready":
-				subMessage = "is not ready"
-			case "Provisioning":
-				subMessage = "is not provisioning"
-			case "Provisioned":
-				subMessage = "is not provisioned"
-			case "Waiting":
-				subMessage = "is waiting"
-			case "Connected":
-				subMessage = "is not connected"
-			case "RKESecretsMigrated":
-				subMessage = "RKE secrets not migrated"
-			case "SecretsMigrated":
-				subMessage = "secrets not migrated"
-			case "NoMemoryPressure":
-				subMessage = "has memory pressure"
-			case "NoDiskPressure":
-				subMessage = "has disk pressure"
-			case "SystemAccountCreated":
-				subMessage = "system account not created"
-			case "SystemProjectCreated":
-				subMessage = "system project not created"
-			case "DefaultProjectCreated":
-				subMessage = "default project not created"
-			case "GlobalAdminsSynced":
-				subMessage = "global admins not synced"
-			case "ServiceAccountMigrated":
-				subMessage = "service account not migrated"
-			case "ServiceAccountSecretsMigrated":
-				subMessage = "service account secrets not migrated"
-			case "AgentDeployed":
-				subMessage = "agent not deployed"
-			case "CreatorMadeOwner":
-				subMessage = "creator not made owner"
-			case "InitialRolesPopulated":
-				subMessage = "initial roles not populated"
-			case "BackingNamespaceCreated":
-				subMessage = "backing namespace not created"
-			}
-			// Add a message for the issue
-			var message string
-			reason := ""
-			msg := ""
-			if len(condition.Reason) > 0 {
-				reason = fmt.Sprintf(", reason is %q", condition.Reason)
-			}
-			if len(condition.Message) > 0 {
-				msg = fmt.Sprintf(", message is %q", condition.Message)
-			}
-			message = fmt.Sprintf("Rancher cluster resource %q, displayed as %s %s%s%s", cluster.Name, cluster.Spec.DisplayName, subMessage, reason, msg)
-			messages = append([]string{message}, messages...)
-		}
-	}
-
-	if len(messages) > 0 {
-		issueReporter.AddKnownIssueMessagesFiles(report.RancherClusterNotReady, clusterRoot, messages, []string{})
-	}
-
-	return nil
+	_ = rancher.AnalyzeClusterRepos(log, clusterRoot, &issueReporter)
+	return rancher.AnalyzeRancherClusters(log, clusterRoot, &issueReporter)
 }
