+++ tools/vz/pkg/analysis/internal/util/cluster/rancher_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/verrazzano/verrazzano/tools/vz/pkg/analysis/internal/util/cluster/rancher"
 	"github.com/verrazzano/verrazzano/tools/vz/pkg/analysis/internal/util/report"
 	"go.uber.org/zap"
 )
@@ -20,14 +21,14 @@ func TestAnalyzeRancher(t *testing.T) {
 
 	// Expect no errors and no reported issues.
 	report.ClearReports()
-	assert.NoError(t, analyzeRancherClusters(logger, "../../../test/cluster/clusters/clusters-ready/cluster-snapshot", &issueReporter))
+	assert.NoError(t, rancher.AnalyzeRancherClusters(logger, "../../../test/cluster/clusters/clusters-ready/cluster-snapshot", &issueReporter))
 	reportedIssues := report.GetAllSourcesFilteredIssues(logger, true, 0, 0)
 	assert.Empty(t, reportedIssues)
 
 	// Expect no errors and one reported issue that a Rancher Cluster is not ready.
 	report.ClearReports()
-	assert.NoError(t, analyzeRancherClusters(logger, "../../../test/cluster/clusters/clusters-not-ready/cluster-snapshot", &issueReporter))
+	assert.NoError(t, rancher.AnalyzeRancherClusters(logger, "../../../test/cluster/clusters/clusters-not-ready/cluster-snapshot", &issueReporter))
 	reportedIssues = report.GetAllSourcesFilteredIssues(logger, true, 0, 0)
 	assert.Len(t, reportedIssues, 1)
-	assert.Equal(t, "RancherClusterNotReady", reportedIssues[0].Type)
+	assert.Equal(t, "RancherIssues", reportedIssues[0].Type)
 }
